% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.consensus_rank}
\alias{autoplot.consensus_rank}
\title{Visualise a consensus ranking}
\usage{
\method{autoplot}{consensus_rank}(object, ...)
}
\arguments{
\item{object}{A \code{consensus_rank} object.}

\item{...}{Reserved for future use.}
}
\value{
A \code{ggplot} object.
}
\description{
Not implemented yet: scheduled for phase F4.
}
