% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{autoplot.judge_clusters}
\alias{autoplot.judge_clusters}
\title{Visualise the disagreement between judges}
\usage{
\method{autoplot}{judge_clusters}(object, ...)
}
\arguments{
\item{object}{A \code{judge_clusters} object.}

\item{...}{Reserved for future use.}
}
\value{
A \code{ggplot} object.
}
\description{
Multidimensional scaling of the judges in the space of the Kemeny-Snell
distance, coloured by cluster.
}
\details{
Not implemented yet: scheduled for phase F4.
}
