% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confsets.R
\name{prob_topk}
\alias{prob_topk}
\title{Probability that a variable lands in the top k}
\usage{
prob_topk(cb, k = 5L)
}
\arguments{
\item{cb}{A \code{rank_confsets} object.}

\item{k}{Size of the top set.}
}
\value{
A tibble of variables and probabilities, in decreasing order.
}
\description{
Not implemented yet: scheduled for phase F3.
}
